update oauth-service for slack workspaces, update tree view for workspaces
