+++ src/manager/treeView.ts
@@ -1,33 +1,34 @@
 import * as vscode from "vscode";
 import {
-  UnreadsTreeProvider,
+  WorkspacesTreeProvider,
   ChannelTreeProvider,
   GroupTreeProvider,
   IMsTreeProvider
 } from "../tree";
 
 export class TreeViewManager implements vscode.Disposable {
-  unreadsTreeProvider: UnreadsTreeProvider;
+  workspacesTreeProvider: WorkspacesTreeProvider;
   channelsTreeProvider: ChannelTreeProvider;
   imsTreeProvider: IMsTreeProvider;
   groupsTreeProvider: GroupTreeProvider;
 
   constructor(public provider: string) {
-    this.unreadsTreeProvider = new UnreadsTreeProvider(provider);
+    this.workspacesTreeProvider = new WorkspacesTreeProvider(provider);
     this.channelsTreeProvider = new ChannelTreeProvider(provider);
     this.groupsTreeProvider = new GroupTreeProvider(provider);
     this.imsTreeProvider = new IMsTreeProvider(provider);
   }
 
-  updateData(channelLabels: ChannelLabel[]) {
-    this.unreadsTreeProvider.update(channelLabels);
-    this.channelsTreeProvider.update(channelLabels);
-    this.groupsTreeProvider.update(channelLabels);
-    this.imsTreeProvider.update(channelLabels);
+  updateData(currentUserInfo: CurrentUser, channelLabels: ChannelLabel[]) {
+    this.workspacesTreeProvider.updateCurrentUser(currentUserInfo);
+    this.workspacesTreeProvider.updateChannels(channelLabels);
+    this.channelsTreeProvider.updateChannels(channelLabels);
+    this.groupsTreeProvider.updateChannels(channelLabels);
+    this.imsTreeProvider.updateChannels(channelLabels);
   }
 
   dispose() {
-    this.unreadsTreeProvider.dispose();
+    this.workspacesTreeProvider.dispose();
     this.channelsTreeProvider.dispose();
     this.groupsTreeProvider.dispose();
     this.imsTreeProvider.dispose();
